feat(application-templates-driving-license): allow only one active application
