+++ libs/application/templates/driving-license/src/dataProviders/index.ts
@@ -1,8 +1,11 @@
-export { QualityPhotoProvider } from '@island.is/application/data-providers'
-export { NationalRegistryProvider } from '@island.is/application/data-providers'
-export { UserProfileProvider } from '@island.is/application/data-providers'
-export { JuristictionProvider } from '@island.is/application/data-providers'
+export {
+  QualityPhotoProvider,
+  NationalRegistryProvider,
+  UserProfileProvider,
+  JuristictionProvider,
+} from '@island.is/application/data-providers'
 export { TeachersProvider } from './TeachersProvider'
 export { DrivingAssessmentProvider } from './DrivingAssessmentProvider'
 export { CurrentLicenseProvider } from './CurrentLicenseProvider'
 export { FeeInfoProvider } from './FeeInfoProvider'
+export { ExistingApplicationProvider } from './ExistingApplicationProvider'
