+++ libs/application/templates/driving-license/src/forms/prerequisites/index.ts
@@ -1,0 +1,1 @@
+export { getForm } from './getForm'
