+++ libs/application/templates/driving-license/src/lib/drivingLicenseTemplate.ts
@@ -5,6 +5,7 @@ import {
   ApplicationStateSchema,
   DefaultStateLifeCycle,
   DefaultEvents,
+  EphemeralStateLifeCycle,
 } from '@island.is/application/core'
 import { FeatureFlagClient } from '@island.is/feature-flags'
 import { ApiActions } from '../shared'
@@ -27,16 +28,13 @@ const template: ApplicationTemplate<
   dataSchema,
   readyForProduction: true,
   stateMachineConfig: {
-    initial: States.DRAFT,
+    initial: States.PREREQUISITES,
     states: {
-      [States.DRAFT]: {
+      [States.PREREQUISITES]: {
         meta: {
           name: m.applicationForDrivingLicense.defaultMessage,
-          actionCard: {
-            description: m.actionCardDraft,
-          },
-          progress: 0.33,
-          lifecycle: DefaultStateLifeCycle,
+          progress: 0.2,
+          lifecycle: EphemeralStateLifeCycle,
           roles: [
             {
               id: Roles.APPLICANT,
@@ -45,17 +43,40 @@ const template: ApplicationTemplate<
                   featureFlagClient as FeatureFlagClient,
                 )
 
-                const getApplication = await import(
-                  '../forms/application'
-                ).then((val) => val.getApplication)
+                const getForm = await import(
+                  '../forms/prerequisites/getForm'
+                ).then((val) => val.getForm)
 
-                return getApplication(
-                  featureFlags[DrivingLicenseFeatureFlags.ALLOW_FAKE],
-                  featureFlags[
-                    DrivingLicenseFeatureFlags.ALLOW_LICENSE_SELECTION
-                  ],
-                )
+                return getForm({
+                  allowFakeData:
+                    featureFlags[DrivingLicenseFeatureFlags.ALLOW_FAKE],
+                  allowPickLicense:
+                    featureFlags[
+                      DrivingLicenseFeatureFlags.ALLOW_LICENSE_SELECTION
+                    ],
+                })
               },
+              write: 'all',
+            },
+          ],
+        },
+        on: {
+          [DefaultEvents.SUBMIT]: { target: States.DRAFT },
+        },
+      },
+      [States.DRAFT]: {
+        meta: {
+          name: m.applicationForDrivingLicense.defaultMessage,
+          actionCard: {
+            description: m.actionCardDraft,
+          },
+          progress: 0.4,
+          lifecycle: DefaultStateLifeCycle,
+          roles: [
+            {
+              id: Roles.APPLICANT,
+              formLoader: async () =>
+                (await import('../forms/application')).application,
               actions: [
                 {
                   event: DefaultEvents.PAYMENT,
@@ -64,6 +85,7 @@ const template: ApplicationTemplate<
                 },
               ],
               write: 'all',
+              read: 'all',
             },
           ],
         },
