+++ libs/application/templates/driving-license/src/lib/utils/hasYes.spec.ts
@@ -1,5 +1,5 @@
 import { Answer } from '@island.is/application/core'
-import { YES, NO } from '../lib/constants'
+import { YES, NO } from '../constants'
 import { hasYes } from './hasYes'
 
 type YesOrNo = typeof YES | typeof NO
