--- libs/application/templates/driving-license/src/utils/index.ts
@@ -1,1 +1,0 @@
-export { hasYes } from './hasYes'
